requirements.txt
python==3.7.1
numpy==1.18.1
pillow==7.2.0
tensorflow==1.15.0
tensorboard==2.0.0
keras==2.3.1
keras-applications==1.0.8
flask==1.1.1
werkzeug==1.0
jinja2==2.11.1
h5py==2.8.0
hdf5==1.8.18
glob2==0.7
opencv-python==4.2.0.32
opencv==4.4.0
pandas==1.0.5
pip==20.1.1
requests==2.24.0
scipy==1.5
yaml==0.1.7

